adding makefile and puzzle maker file
Added targets for make and make clean for when the project needs to be
built and when the created files need to be removed.

Added the primary file where the bulk of the logic for the Picross
Puzzle Generator will go.

This is just a primary commit, ostensibly a 'first commit' for file
setup and such.