adding sequence counter to calculate hint numbers
An integral part of the Picross puzzle game are the hint numbers. These
numbers let the player know about contiguous 'true' or 'filled'
sequences within a row or column. In order to adhere to that
functionality, a contiguous sequence counter has been added for both
rows and columns. This will track hint numbers in arrays respective to
their particular dimensions.

This is an initial and sloppy implementation. Future enhancements will
be made. Encapsulation will be considered (weighing the pros and cons of
such a method). Asynchronous array traversal will also be considered.
Finally, if possible it would be wonderful to remove any column-major
array traversal in order to fully optimize performance (not a necessity,
but it is a nice hands-on learning tool for performance focused
methodologies).